refactor: :recycle: move getActiveTaxProvider to active-tax-provider
