+++ apps/taxes/src/pages/api/webhooks/order-calculate-taxes.ts
@@ -3,7 +3,7 @@ import { UntypedCalculateTaxesDocument } from "../../../../generated/graphql";
 import { saleorApp } from "../../../../saleor-app";
 import { createLogger } from "../../../lib/logger";
 import { calculateTaxesPayloadSchema, ExpectedWebhookPayload } from "../../../lib/saleor/schema";
-import { ActiveTaxProviderService } from "../../../modules/taxes/active-tax-provider.service";
+import { getActiveTaxProvider } from "../../../modules/taxes/active-tax-provider";
 
 export const config = {
   api: {
@@ -38,9 +38,7 @@ export default orderCalculateTaxesSyncWebhook.createHandler(async (req, res, ctx
   try {
     const appMetadata = data.recipient?.privateMetadata ?? [];
     const channelSlug = data.taxBase.channel.slug;
-
-    const activeTaxProviderService = new ActiveTaxProviderService();
-    const taxProvider = await activeTaxProviderService.get(channelSlug, appMetadata);
+    const taxProvider = getActiveTaxProvider(channelSlug, appMetadata);
     logger.info({ taxProvider }, "Fetched activeTaxProvider");
 
     const calculatedTaxes = await taxProvider.calculateTaxes(data.taxBase);
