+++ apps/taxes/src/pages/api/webhooks/order-fulfilled.ts
@@ -5,8 +5,7 @@ import {
 } from "../../../../generated/graphql";
 import { saleorApp } from "../../../../saleor-app";
 import { createLogger } from "../../../lib/logger";
-import { ActiveTaxProviderService } from "../../../modules/taxes/active-tax-provider.service";
-
+import { getActiveTaxProvider } from "../../../modules/taxes/active-tax-provider";
 export const config = {
   api: {
     bodyParser: false,
@@ -34,9 +33,7 @@ export default orderFulfilledAsyncWebhook.createHandler(async (req, res, ctx) =>
   try {
     const appMetadata = payload.recipient?.privateMetadata ?? [];
     const channelSlug = payload.order?.channel.slug;
-
-    const activeTaxProviderService = new ActiveTaxProviderService();
-    const taxProvider = await activeTaxProviderService.get(channelSlug, appMetadata);
+    const taxProvider = getActiveTaxProvider(channelSlug, appMetadata);
     logger.info({ taxProvider }, "Fetched activeTaxProvider");
 
     // todo: figure out what fields are needed and add validation
